+++ StreamDeck-KeePass/StreamDeck-KeePass/PropertyInspector/KeePassRetrievePI.html
@@ -24,8 +24,8 @@
             <input class="sdpi-item-value sdProperty" placeholder="The password for the .kdbx" value="" id="Password" oninput="setSettings()" type="password">
         </div>
         <div class="sdpi-item" id="dvEntryTitle">
-            <div class="sdpi-item-label">Entry Name</div>
-            <input class="sdpi-item-value sdProperty" placeholder="The name of the KeePass entry to retrieve the password" value="" id="entryTitle" oninput="setSettings()">
+            <div class="sdpi-item-label">Entry Name or UUID</div>
+            <input class="sdpi-item-value sdProperty" placeholder="The name or UUID of the KeePass entry to retrieve the password" value="" id="entryTitle" oninput="setSettings()">
         </div>
         <div class="sdpi-item" id="dvField">
             <div class="sdpi-item-label">Select field to retrieve</div>
